requirements.txt
<<<<<<< HEAD
clip
pyfiglet
Pillow
=======
clip==1.0
pyfiglet==0.8.post1
Pillow==8.2.0
>>>>>>> first commit , add basic files
